Add tantan api
