+++ service/chat/adapter/devices.go
