+++ service/chat/app/application.go
@@ -4,7 +4,7 @@ import (
 	"fxlibraries/httpserver"
 	"fxlibraries/loggers"
 
-	"fxservice/service/momo/handlers"
+	"fxservice/service/chat/handlers"
 )
 
 func init() {
@@ -15,16 +15,16 @@ func Start(addr string) {
 	r := httpserver.NewRouter()
 	r.RouteHandleFunc("/status", handlers.Test).Methods("GET")
 
-	r.RouteHandleFunc("/accounts", handlers.GetFreeAccounts).Methods("GET").Queries("action", "online")
+	r.RouteHandleFunc("/accounts", handlers.GetFreeMomoAccounts).Methods("GET").Queries("action", "online")
 	r.RouteHandleFunc("/accounts", handlers.GetMomoAccounts).Methods("GET")
 	r.RouteHandleFunc("/accounts", handlers.AddAccounts).Methods("POST")
-	r.RouteHandleFunc("/accounts", handlers.GetFreeAccounts).Methods("PATCH").Queries("action", "online")
+	r.RouteHandleFunc("/accounts", handlers.GetFreeMomoAccounts).Methods("PATCH").Queries("action", "online")
 	r.RouteHandleFunc("/accounts", handlers.PatchMomoAccounts).Methods("PATCH")
 
 	r.RouteHandleFunc("/accounts/new", handlers.UnRegisterMomoAccounts).Methods("GET")
 	r.RouteHandleFunc("/accounts/{account}", handlers.CompleteMomoAccount).Methods("PATCH")
 
-	r.RouteHandleFunc("/replys/{account}", handlers.GetAccountReply).Methods("GET")
+	r.RouteHandleFunc("/replys/{account}", handlers.GetMomoAccountReply).Methods("GET")
 
 	r.RouteHandleFunc("/gpss", handlers.AddGPSs).Methods("POST")
 	r.RouteHandleFunc("/photos", handlers.AddPhotos).Methods("POST")
