+++ service/chat/common/devices.go
