+++ service/chat/common/model.go
