+++ service/chat/config/config.go
@@ -21,7 +21,7 @@ var (
 )
 
 var (
-	configFilePath = flag.String("c", "/etc/fxservice/momo/config.json", "config file path")
+	configFilePath = flag.String("c", "/etc/fxservice/chat/config.json", "config file path")
 	isDebug        = flag.Bool("d", false, "debug mode")
 	isLocal        = flag.Bool("local", false, "local develop mode")
 )
