+++ service/chat/handlers/accounts.go
@@ -5,12 +5,57 @@ import (
 	"fxlibraries/httpserver"
 	"fxlibraries/loggers"
 	"fxservice/domain"
-	"fxservice/service/momo/adapter"
-	"fxservice/service/momo/common"
+	"fxservice/service/chat/adapter"
+	"fxservice/service/chat/common"
 	"strconv"
 	"strings"
 	"time"
 )
+
+func AddAccounts(req *httpserver.Request) *httpserver.Response {
+	var accounts []domain.MomoAccount
+	if err := req.Parse(&accounts); err != nil {
+		loggers.Warn.Printf("AddAccounts parse accounts error %s", err.Error())
+		return httpserver.NewResponseWithError(errors.ParameterError)
+	}
+	var newAccounts []domain.MomoAccount
+	for i := range accounts {
+		if accounts[i].Account == "" || accounts[i].Password == "" {
+			loggers.Warn.Printf("AddAccounts invalid account %s:%s", accounts[i].Account, accounts[i].Password)
+			continue
+		}
+		if accounts[i].AccountType == 0 {
+			accounts[i].AccountType = domain.QQ
+		}
+		device, err := adapter.GetEnableDevice()
+		if err != nil {
+			loggers.Warn.Printf("AddAccounts get enable device error %s", err.Error())
+			return httpserver.NewResponseWithError(errors.NewNotFound("No enable devices"))
+		}
+		nickName, err := adapter.GetRandNickName()
+		if err != nil {
+			loggers.Warn.Printf("AddAccounts get nickname error %s", err.Error())
+			return httpserver.NewResponseWithError(errors.NewNotFound("No nicknames "))
+		}
+
+		operator := common.GenRandOperator()
+		accounts[i].NickName = nickName.NickName
+		accounts[i].MomoPassword = common.GenRandPassword(8)
+		accounts[i].Operator = operator.Operator
+		accounts[i].OperatorMC = operator.OperatorMC
+		accounts[i].OperatorMN = operator.OperatorMN
+		accounts[i].SN = device.SN
+		accounts[i].Status = domain.MomoAccountStatusUnRegister
+		accounts[i].Gender = domain.Female
+		newAccounts = append(newAccounts, accounts[i])
+	}
+	if err := adapter.AddAccounts(&newAccounts); err != nil {
+		loggers.Warn.Printf("AddAccounts error %s", err.Error())
+		return httpserver.NewResponseWithError(errors.InternalServerError)
+	}
+
+	return httpserver.NewResponse()
+}
 
 func UnRegisterMomoAccounts(req *httpserver.Request) *httpserver.Response {
 	province := req.QueryParams.Get("province")
@@ -78,51 +123,6 @@ func UnRegisterMomoAccounts(req *httpserver.Request) *httpserver.Response {
 
 	loggers.Warn.Println(resp)
 	return resp
-}
-
-func AddAccounts(req *httpserver.Request) *httpserver.Response {
-	var accounts []domain.MomoAccount
-	if err := req.Parse(&accounts); err != nil {
-		loggers.Warn.Printf("AddAccounts parse accounts error %s", err.Error())
-		return httpserver.NewResponseWithError(errors.ParameterError)
-	}
-	var newAccounts []domain.MomoAccount
-	for i := range accounts {
-		if accounts[i].Account == "" || accounts[i].Password == "" {
-			loggers.Warn.Printf("AddAccounts invalid account %s:%s", accounts[i].Account, accounts[i].Password)
-			continue
-		}
-		if accounts[i].AccountType == 0 {
-			accounts[i].AccountType = domain.QQ
-		}
-		device, err := adapter.GetEnableDevice()
-		if err != nil {
-			loggers.Warn.Printf("AddAccounts get enable device error %s", err.Error())
-			return httpserver.NewResponseWithError(errors.NewNotFound("No enable devices"))
-		}
-		nickName, err := adapter.GetRandNickName()
-		if err != nil {
-			loggers.Warn.Printf("AddAccounts get nickname error %s", err.Error())
-			return httpserver.NewResponseWithError(errors.NewNotFound("No nicknames "))
-		}
-
-		operator := common.GenRandOperator()
-		accounts[i].NickName = nickName.NickName
-		accounts[i].MomoPassword = common.GenRandPassword(8)
-		accounts[i].Operator = operator.Operator
-		accounts[i].OperatorMC = operator.OperatorMC
-		accounts[i].OperatorMN = operator.OperatorMN
-		accounts[i].SN = device.SN
-		accounts[i].Status = domain.MomoAccountStatusUnRegister
-		accounts[i].Gender = domain.Female
-		newAccounts = append(newAccounts, accounts[i])
-	}
-	if err := adapter.AddAccounts(&newAccounts); err != nil {
-		loggers.Warn.Printf("AddAccounts error %s", err.Error())
-		return httpserver.NewResponseWithError(errors.InternalServerError)
-	}
-
-	return httpserver.NewResponse()
 }
 
 func CompleteMomoAccount(req *httpserver.Request) *httpserver.Response {
@@ -246,7 +246,7 @@ func PatchMomoAccounts(req *httpserver.Request) *httpserver.Response {
 	return httpserver.NewResponse()
 }
 
-func GetFreeAccounts(req *httpserver.Request) *httpserver.Response {
+func GetFreeMomoAccounts(req *httpserver.Request) *httpserver.Response {
 	province := req.QueryParams.Get("province")
 	city := req.QueryParams.Get("city")
 	account := req.QueryParams.Get("account")
@@ -318,7 +318,7 @@ func GetFreeAccounts(req *httpserver.Request) *httpserver.Response {
 
 }
 
-func GetAccountReply(req *httpserver.Request) *httpserver.Response {
+func GetMomoAccountReply(req *httpserver.Request) *httpserver.Response {
 	account := req.UrlParams["account"]
 	if account == "" {
 		loggers.Warn.Printf("GetAccountReply no account")
