+++ service/chat/script/device_upload.py
